major reorganization
option to use database checks instead of LLM for free living control proteome selection
