+++ src/taxon_selection/fetch_proteomes.sh
